+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2023-03-25 14:10:04
+
+- [x] update visual reasoning query
+
 
 ## 2022-10-27 12:01:37